[TASK] Fix phpstan checkFunctionArgumentTypes errors in ext:core Localization
This patch fixes incompatible type usage in function arguments
and is preparatory work for introducing native type hints and
strict mode in all core files.

Releases: master, 10.4
Resolves: #92266
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65729
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>

+++ typo3/sysext/core/Classes/Localization/Parser/XliffParser.php
@@ -77,7 +77,7 @@ class XliffParser extends AbstractXmlParser
                             $id = (string)$translationElement['id'];
                         } else {
                             $id = (string)$translationElement->{'trans-unit'}[0]['id'];
-                            $id = substr($id, 0, strpos($id, '['));
+                            $id = substr($id, 0, (int)strpos($id, '['));
                         }
                         $parsedData[$id] = $parsedTranslationElement;
                     }
